requirements.txt
editor
safer
pyyaml
versy
requests
beautifulsoup4
